- Moved the Scanning classes to their own package
