+++ src/burp/ParamGuesser.java
@@ -1,14 +1,14 @@
 package burp;
 
 import burp.albinowaxUtils.Attack;
-import burp.albinowaxUtils.BulkScan;
+import burp.model.scanning.BulkScan;
 import burp.view.ConfigurableSettings;
 import burp.albinowaxUtils.CustomScanIssue;
 import burp.albinowaxUtils.ParamInsertionPoint;
-import burp.albinowaxUtils.ParamScan;
+import burp.model.scanning.ParamScan;
 import burp.albinowaxUtils.PayloadInjector;
 import burp.albinowaxUtils.RawInsertionPoint;
-import burp.albinowaxUtils.Scan;
+import burp.model.scanning.Scan;
 import org.apache.commons.collections4.queue.CircularFifoQueue;
 
 import java.util.*;
